Feat: added index tab to load necessary data on sign in.
